Core: Move EXI source files to their own directory
Keeps related source files together and cleans up the root directory of HW
a little bit.
