+++ Source/Core/Core/Boot/Boot_BS2Emu.cpp
@@ -15,7 +15,7 @@
 #include "Core/HLE/HLE.h"
 #include "Core/HW/CPU.h"
 #include "Core/HW/DVDInterface.h"
-#include "Core/HW/EXI_DeviceIPL.h"
+#include "Core/HW/EXI/EXI_DeviceIPL.h"
 #include "Core/HW/Memmap.h"
 #include "Core/MemTools.h"
 #include "Core/PatchEngine.h"
