+++ Source/Core/Core/ConfigManager.h
@@ -10,7 +10,7 @@
 
 #include "Common/IniFile.h"
 #include "Common/NonCopyable.h"
-#include "Core/HW/EXI_Device.h"
+#include "Core/HW/EXI/EXI_Device.h"
 #include "Core/HW/SI_Device.h"
 
 namespace DiscIO
