+++ Source/Core/Core/Core.vcxproj
@@ -96,7 +96,6 @@
     <ClCompile Include="HLE\HLE_OS.cpp" />
     <ClCompile Include="HotkeyManager.cpp" />
     <ClCompile Include="HW\AudioInterface.cpp" />
-    <ClCompile Include="HW\BBA-TAP\TAP_Win32.cpp" />
     <ClCompile Include="HW\CPU.cpp" />
     <ClCompile Include="HW\DSP.cpp" />
     <ClCompile Include="HW\DSPHLE\DSPHLE.cpp" />
@@ -117,16 +116,17 @@
     <ClCompile Include="HW\DSPLLE\DSPSymbols.cpp" />
     <ClCompile Include="HW\DVDInterface.cpp" />
     <ClCompile Include="HW\DVDThread.cpp" />
-    <ClCompile Include="HW\EXI.cpp" />
-    <ClCompile Include="HW\EXI_Channel.cpp" />
-    <ClCompile Include="HW\EXI_Device.cpp" />
-    <ClCompile Include="HW\EXI_DeviceAD16.cpp" />
-    <ClCompile Include="HW\EXI_DeviceAGP.cpp" />
-    <ClCompile Include="HW\EXI_DeviceEthernet.cpp" />
-    <ClCompile Include="HW\EXI_DeviceGecko.cpp" />
-    <ClCompile Include="HW\EXI_DeviceIPL.cpp" />
-    <ClCompile Include="HW\EXI_DeviceMemoryCard.cpp" />
-    <ClCompile Include="HW\EXI_DeviceMic.cpp" />
+    <ClCompile Include="HW\EXI\BBA-TAP\TAP_Win32.cpp" />
+    <ClCompile Include="HW\EXI\EXI.cpp" />
+    <ClCompile Include="HW\EXI\EXI_Channel.cpp" />
+    <ClCompile Include="HW\EXI\EXI_Device.cpp" />
+    <ClCompile Include="HW\EXI\EXI_DeviceAD16.cpp" />
+    <ClCompile Include="HW\EXI\EXI_DeviceAGP.cpp" />
+    <ClCompile Include="HW\EXI\EXI_DeviceEthernet.cpp" />
+    <ClCompile Include="HW\EXI\EXI_DeviceGecko.cpp" />
+    <ClCompile Include="HW\EXI\EXI_DeviceIPL.cpp" />
+    <ClCompile Include="HW\EXI\EXI_DeviceMemoryCard.cpp" />
+    <ClCompile Include="HW\EXI\EXI_DeviceMic.cpp" />
     <ClCompile Include="HW\GCKeyboard.cpp" />
     <ClCompile Include="HW\GCKeyboardEmu.cpp" />
     <ClCompile Include="HW\GCMemcard.cpp" />
@@ -322,7 +322,6 @@
     <ClInclude Include="Host.h" />
     <ClInclude Include="HotkeyManager.h" />
     <ClInclude Include="HW\AudioInterface.h" />
-    <ClInclude Include="HW\BBA-TAP\TAP_Win32.h" />
     <ClInclude Include="HW\CPU.h" />
     <ClInclude Include="HW\DSP.h" />
     <ClInclude Include="HW\DSPHLE\DSPHLE.h" />
@@ -344,16 +343,17 @@
     <ClInclude Include="HW\DSPLLE\DSPSymbols.h" />
     <ClInclude Include="HW\DVDInterface.h" />
     <ClInclude Include="HW\DVDThread.h" />
-    <ClInclude Include="HW\EXI.h" />
-    <ClInclude Include="HW\EXI_Channel.h" />
-    <ClInclude Include="HW\EXI_Device.h" />
-    <ClInclude Include="HW\EXI_DeviceAD16.h" />
-    <ClInclude Include="HW\EXI_DeviceAGP.h" />
-    <ClInclude Include="HW\EXI_DeviceEthernet.h" />
-    <ClInclude Include="HW\EXI_DeviceGecko.h" />
-    <ClInclude Include="HW\EXI_DeviceIPL.h" />
-    <ClInclude Include="HW\EXI_DeviceMemoryCard.h" />
-    <ClInclude Include="HW\EXI_DeviceMic.h" />
+    <ClInclude Include="HW\EXI\BBA-TAP\TAP_Win32.h" />
+    <ClInclude Include="HW\EXI\EXI.h" />
+    <ClInclude Include="HW\EXI\EXI_Channel.h" />
+    <ClInclude Include="HW\EXI\EXI_Device.h" />
+    <ClInclude Include="HW\EXI\EXI_DeviceAD16.h" />
+    <ClInclude Include="HW\EXI\EXI_DeviceAGP.h" />
+    <ClInclude Include="HW\EXI\EXI_DeviceEthernet.h" />
+    <ClInclude Include="HW\EXI\EXI_DeviceGecko.h" />
+    <ClInclude Include="HW\EXI\EXI_DeviceIPL.h" />
+    <ClInclude Include="HW\EXI\EXI_DeviceMemoryCard.h" />
+    <ClInclude Include="HW\EXI\EXI_DeviceMic.h" />
     <ClInclude Include="HW\GCKeyboard.h" />
     <ClInclude Include="HW\GCKeyboardEmu.h" />
     <ClInclude Include="HW\GCMemcard.h" />
