+++ Source/Core/Core/HW/EXI/BBA-TAP/TAP_Apple.cpp
@@ -7,8 +7,8 @@
 
 #include "Common/Logging/Log.h"
 #include "Common/StringUtil.h"
-#include "Core/HW/EXI_Device.h"
-#include "Core/HW/EXI_DeviceEthernet.h"
+#include "Core/HW/EXI/EXI_Device.h"
+#include "Core/HW/EXI/EXI_DeviceEthernet.h"
 
 bool CEXIETHERNET::Activate()
 {
