+++ Source/Core/Core/HW/EXI/EXI_DeviceAD16.cpp
@@ -2,7 +2,8 @@
 // Licensed under GPLv2+
 // Refer to the license.txt file included.
 
-#include "Core/HW/EXI_DeviceAD16.h"
+#include "Core/HW/EXI/EXI_DeviceAD16.h"
+
 #include "Common/Assert.h"
 #include "Common/ChunkFile.h"
 #include "Common/CommonTypes.h"
