+++ Source/Core/Core/HW/EXI/EXI_DeviceAD16.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "Core/HW/EXI_Device.h"
+#include "Core/HW/EXI/EXI_Device.h"
 
 class PointerWrap;
