+++ Source/Core/Core/HW/EXI/EXI_DeviceEthernet.h
@@ -13,7 +13,7 @@
 #endif
 
 #include "Common/Flag.h"
-#include "Core/HW/EXI_Device.h"
+#include "Core/HW/EXI/EXI_Device.h"
 
 class PointerWrap;
