+++ Source/Core/Core/HW/EXI/EXI_DeviceIPL.h
@@ -6,7 +6,7 @@
 
 #include <string>
 
-#include "Core/HW/EXI_Device.h"
+#include "Core/HW/EXI/EXI_Device.h"
 
 class PointerWrap;
