+++ Source/Core/Core/HW/EXI/EXI_DeviceMemoryCard.h
@@ -7,7 +7,7 @@
 #include <functional>
 #include <memory>
 
-#include "Core/HW/EXI_Device.h"
+#include "Core/HW/EXI/EXI_Device.h"
 
 class MemoryCardBase;
 class PointerWrap;
